lib: Rename "path" to "cockpit-path"
To make it clear that this is our module.
