+++ pkg/lib/cockpit-path.ts
